chore: implement multi-platform CI/CD workflows, modernize portal layout structures, and refactor cross-app native components and service layers.

+++ pegasusX/apps/driver-app-android/app/src/main/java/com/pegasusx/driver/ui/screens/offload/components/OffloadSummaryCard.kt
@@ -21,8 +21,8 @@ import com.pegasusx.driver.ui.theme.formattedAmount
 
 @Composable
 fun OffloadSummaryCard(
-    originalTotal: Double,
-    adjustedTotal: Double,
+    originalTotal: Long,
+    adjustedTotal: Long,
     hasRejections: Boolean
 ) {
     val lab = LocalPegasusColors.current
